Model & Client Audit (#43)
* Switch to same enum style list call as Images

* Should have been imageIdOrSlug all along. imageId being an int and slug being a string.

* Rename ForwardingRules->ForwardingRule.
Add ForwardingRulesList, the list request used in adding/removing.
Explicit ForwardRule request (1 optional value type) and response class.
LoadBalancer request modified, 2 optional value types.

* Match filename to class name

* Rename Address to IpAddress (same as Networks and DomainRecords)
Forgot to removed un-necessary JsonPropertAttributes from response

* Create explicit HealthCheck request class (4 optional value type), remove un-necessary JsonPropertyAttributes from response class.
Rename type->Type

* Change currency properties from float to Decimal type

* Create Explicit UpdateImage class, convert Image to actual class for Create method.
Move ImageType enum to its own file.
Simplify GetAll switch logic.

* Explicit UpdateKey class (can only update name currently).
There is no UpdateFingerprint to Test.

* Tags are always strings.
4 optional value type properties.

* Rename request TagResource->TagResources, strongly typed request class for Tag/Untag methods.
Renamed request Resource->TagResource and move into its own file.
Rename Resource->TagResourcesStatistics, strongly typed respone class. Add missing props. Created Base TagResourceStatistics class as well.

* Rename to ForwardingRules, the methods operate on multiple items.

* Explicit UpdateDomainRecord class.

* Explicit Requests class for StickySessions.

* LoadBalancer id's are strings not ints

* DomainRecord doc updates

* Forgot [Fact]

* Consistent usage of List over array.

* Capitalization

* Doc update

* Should be a floating point, defaulting to double.

* Separate V4 and V6 Interface classes.
V6 uses integer for Netmask.

* Split Action request into DropletAction and ImageAction.

+++ DigitalOcean.API/Models/Responses/Image.cs
@@ -53,8 +53,8 @@ namespace DigitalOcean.API.Models.Responses {
         public int MinDiskSize { get; set; }
 
         /// <summary>
-        /// The size of the image in gigabytes
+        /// The size of the image in gigabytes.
         /// </summary>
-        public int SizeGigabytes { get; set; }
+        public double SizeGigabytes { get; set; }
     }
 }
